http: extract worker into a separate compilation unit
