+++ modules/http/Makefile.am
@@ -4,6 +4,8 @@ module_LTLIBRARIES      += modules/http/libhttp.la
 modules_http_libhttp_la_SOURCES = \
   modules/http/http.h 		    \
   modules/http/http.c               \
+  modules/http/http-worker.c	    \
+  modules/http/http-worker.h	    \
   modules/http/http-grammar.y       \
   modules/http/http-parser.c        \
   modules/http/http-parser.h        \
