+++ modules/http/http.h
@@ -30,17 +30,6 @@
 
 #include "logthrdestdrv.h"
 
-#define CURL_NO_OLDIES 1
-#include <curl/curl.h>
-
-typedef struct _HTTPDestinationWorker
-{
-  LogThreadedDestWorker super;
-  CURL *curl;
-  GString *request_body;
-  struct curl_slist *request_headers;
-} HTTPDestinationWorker;
-
 typedef struct
 {
   LogThreadedDestDriver super;
@@ -93,9 +82,5 @@ void http_dd_set_body_prefix(LogDriver *d, const gchar *body_prefix);
 void http_dd_set_body_suffix(LogDriver *d, const gchar *body_suffix);
 void http_dd_set_delimiter(LogDriver *d, const gchar *delimiter);
 LogTemplateOptions *http_dd_get_template_options(LogDriver *d);
-
-/* internal api */
-worker_insert_result_t map_http_status_to_worker_status(HTTPDestinationWorker *self, glong http_code);
-HTTPDestinationWorker *http_dw_new(HTTPDestinationDriver *owner, gint worker_index);
 
 #endif
